Initial tz creation from VTIMEZONE components
This is currently brute-forced into cal.py, and only works when
VTIMEZONE components come before DATE-TIME components that make use of
them (while this seems to be the standard way to generate icalendars,
there is no word about this in the RFC). To circumvent this, we would
probably need two rounds of parsing.

There are so many exceptions in from_ical() already, that I don't feel
too guilty about introducing one more.

Tests are needed.