parser: Ignore UTF-8 BOM characters
The zero width No-Break space U+FEFF (also known as BOM) is not really
supposed to occur in ics files. However some particularly bad ics file
writers generate calendars with a BOM at the beginning.

Decoding with the encoding set to "utf-8-sig" is the easiest way to deal
with this [0], particularly because:
- If the file is BOM encoded UTF-8, then "utf-8" leaves the BOM in the
  content, while "utf-8-sig" transparently removes the BOM
- If the file is Standard UTF-8, then "utf-8-sig" decodes the same as
  "utf-8"

[0] https://stackoverflow.com/a/44573867